% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_doc.R
\name{theme_doc}
\alias{theme_doc}
\title{This is a redesigned ggplot theme built for out of the box scientific publication
Use this theme with your normal ggplot function}
\usage{
theme_doc(base_size = 14, base_family = "sans")
}
\description{
This is a redesigned ggplot theme built for out of the box scientific publication
Use this theme with your normal ggplot function
}
