% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_doc.R
\name{theme_doc}
\alias{theme_doc}
\title{ggplot doc theme}
\usage{
theme_doc(base_size = 14, base_family = "sans", grid = FALSE)
}
\arguments{
\item{base_size}{Base font size. Default 14}

\item{base_family}{Base font family. Default "sans"}

\item{grid}{Boolean switch to turn major grid on or off. Default FALSE}
}
\description{
This is a redesigned ggplot theme built for creating out of the box figures for use in scientific publications.
}
\author{
Rishav Ray \email{rishav.rray@gmail.com}
}
